pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "money3"
version = "0.1.0"
description = "基于 Black-Litterman 模型的智能投资组合优化系统"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "matplotlib>=3.8",
    "PyPortfolioOpt>=1.5.5",
    "vectorbt>=0.25.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "openai>=1.51.0",
    "finnhub-python>=2.4.20",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
include = ["money3*"]
